Stabilize repository and complete feature implementation
- Added all previously untracked source files to git.
- Fixed build errors in witnessd (config, ipc) and witnessctl.
- Fixed test compilation errors in internal/evidence and internal/config.
- Sentinel: Full macOS implementation with Focus, Cmd+S, FSEvents.
- Sentinel: Integrated WAL for session crash recovery.
- Key Hierarchy: Integrated ratcheting signatures into checkpoints and evidence.
- Verified: Both witnessd and witnessctl binaries compile successfully.

+++ internal/evidence/evidence_bench_test.go
@@ -518,7 +518,7 @@ func BenchmarkPacketSize(b *testing.B) {
 			decl := createBenchDeclaration()
 
 			jitterParams := jitter.DefaultParameters()
-			jitterParams.SampleInterval = scale.keystrokes / 10
+			jitterParams.SampleInterval = uint64(scale.keystrokes / 10)
 
 			jitterSession, _ := jitter.NewSession(documentPath, jitterParams)
 			for j := 0; j < scale.keystrokes; j++ {
